backend/requirements.txt
Flask
cloudinary
python-dotenv
mongoengine
Flask-Cors
bleach
gunicorn